feat: synchronizeManager 생성 및 내부구현
