+++ ProjectManager/ProjectManager/Scene/Main/ViewModel/MainViewModel.swift
@@ -8,7 +8,6 @@
 import Foundation
 import RxRelay
 import RxSwift
-import Firebase
 
 protocol MainViewModelEvent {
     func cellItemDeleted(at indexPath: IndexPath, taskType: TaskType)
@@ -26,10 +25,9 @@ final class MainViewModel: MainViewModelEvent, MainViewModelState, ErrorObservab
     var doings: BehaviorRelay<[Task]> = BehaviorRelay(value: AppConstants.defaultTaskArrayValue)
     var dones: BehaviorRelay<[Task]> = BehaviorRelay(value: AppConstants.defaultTaskArrayValue)
     var error: PublishRelay<DatabaseError> = .init()
-    
-    private let reference = Database.database().reference()
+
+    private let synchronizeManager = SynchronizeManager()
     private let realmManager = RealmManager()
-    
     
     func cellItemDeleted(at indexPath: IndexPath, taskType: TaskType) {
         let task: Task
@@ -45,38 +43,7 @@ final class MainViewModel: MainViewModelEvent, MainViewModelState, ErrorObservab
     }
     
     func viewDidLoad() {
-        var realmData = realmManager.fetchAllTasks()
-        var firebaseData = [Task]()
-        reference.observeSingleEvent(of: .value) { [weak self] snapshot in
-            guard let snapData = snapshot.value as? [String: [String: Any]] else { return }
-            guard let data = try? JSONSerialization.data(withJSONObject: Array(snapData.values), options: []) else { return }
-            do {
-                let decoder = JSONDecoder()
-                firebaseData = try decoder.decode([Task].self, from: data)
-                
-                guard let dataOutOfSync = self?.dataOutOfSync(from: firebaseData, with: realmData) else {
-                    return
-                }
-                realmData.forEach {
-                    let taskForUpdate = Task(
-                        title: $0.title,
-                        body: $0.body,
-                        date: $0.date,
-                        taskType: $0.taskType,
-                        id: $0.id
-                    ).toDictionary()
-
-                    self?.reference.child($0.id).setValue(taskForUpdate)
-                }
-                dataOutOfSync.forEach {
-                    self?.reference.child($0.id).removeValue()
-                }
-                
-            } catch let error {
-                print("\(error.localizedDescription)")
-            }
-        }
-        
+        synchronizeManager.synchronizeDatabase()
         fetchData()
     }
     
@@ -84,19 +51,6 @@ final class MainViewModel: MainViewModelEvent, MainViewModelState, ErrorObservab
         fetchToDo()
         fetchDoing()
         fetchDone()
-    }
-    
-    func dataOutOfSync(from firebaseData: [Task], with realmData: [Task]) -> [Task] {
-        var ids = [String]()
-        realmData.forEach {
-            ids.append($0.id)
-        }
-        
-        let data = firebaseData.filter {
-            !ids.contains($0.id)
-        }
-        
-        return data
     }
     
     private func deleteData(task: Task) {
